requirements.txt
isort==5.13.2
mypy==1.14.0
pre-commit==4.0.1
ruff==0.8.4
telethon>=1.28.5
aiohttp>=3.9.0
fastapi>=0.110.0
uvicorn>=0.27.1
pydantic>=2.6.3
python-dotenv==1.0.0